Seperate filter_vcfs from create_cbio_study file
