+++ filter_vcfs.py
@@ -1,0 +1,53 @@
+import gzip
+
+
+def filter_vcfs(sample_id, museq_vcf, strelka_vcf, work_dir):
+    '''
+    original code by Diljot Grewal
+
+    museq_paired and strekla_snv,
+    take position intersection plus probability filter of 0.85
+    (keep positions >= 0.85 that are in both)
+
+    modifications: takes museq and strelka directly as input, outputs
+    to temp_dir, takes sample id as input and append to output
+    filtered filename
+    '''
+
+    museq_filtered = work_dir + '{}_museq_filtered.vcf.gz'.format(sample_id)
+
+    strelka_ref = set()
+
+    with gzip.open(strelka_vcf, 'rt') as strelka_data:
+        for line in strelka_data:
+            if line.startswith('#'):
+                continue
+            
+            line = line.strip().split()
+            
+            chrom = line[0]
+            pos = line[1]
+            
+            strelka_ref.add((chrom, pos))
+
+    with gzip.open(museq_vcf, 'rt') as museq_data, gzip.open(museq_filtered, 'wt') as museqout:
+        for line in museq_data:
+            if line.startswith('#'):
+                museqout.write(line)
+                continue
+            
+            line = line.strip().split()
+            chrom = line[0]
+            pos = line[1]
+            
+            if ((chrom, pos))  not in strelka_ref:
+                continue
+            
+            pr = line[7].split(';')[0].split('=')[1]
+            if float(pr) < 0.85:
+                continue
+            
+            outstr = '\t'.join(line)+'\n'
+            museqout.write(outstr)
+
+    return museq_filtered
